Move common utils to test-support
