+++ tests/integration/components/frost-bunsen-array-inline-item-test.js
@@ -1,6 +1,6 @@
 import {expect} from 'chai'
 import {beforeEach, describe, it} from 'mocha'
-import {setupComponentTest} from '../../utils/template'
+import {setupComponentTest} from 'dummy/tests/helpers/template'
 
 const props = {
   bunsenId: 'addresses',
