+++ tests/integration/components/frost-bunsen-cell-test.js
@@ -3,7 +3,7 @@ import {it} from 'ember-mocha'
 import {beforeEach, describe} from 'mocha'
 import {recursiveObjectCreate} from 'ember-frost-bunsen/utils'
 import {builtInRenderers} from 'ember-frost-bunsen/validator/index'
-import {setupComponentTest} from '../../utils/template'
+import {setupComponentTest} from 'dummy/tests/helpers/template'
 
 import model from '../../fixtures/valid-model'
 import view from '../../fixtures/valid-view'
