+++ tests/integration/components/frost-bunsen-container-test.js
@@ -1,6 +1,6 @@
 import {expect} from 'chai'
 import {it} from 'ember-mocha'
-import {setupComponentTest} from '../../utils/template'
+import {setupComponentTest} from 'dummy/tests/helpers/template'
 
 const props = {
   cellConfig: Ember.Object.create({}),
