+++ tests/unit/components/array-tab-content-test.js
@@ -3,7 +3,7 @@ import {describeComponent} from 'ember-mocha'
 import {afterEach, beforeEach, it} from 'mocha'
 import {PropTypes} from 'ember-prop-types'
 import {builtInRenderers} from 'ember-frost-bunsen/validator/index'
-import {validatePropTypes} from '../../utils/template'
+import {validatePropTypes} from 'dummy/tests/helpers/template'
 
 describeComponent(
   'frost-bunsen-array-tab-content',
