+++ tests/unit/components/button-group-input-test.js
@@ -3,7 +3,8 @@ import {describeComponent, it} from 'ember-mocha'
 import {afterEach, beforeEach, describe} from 'mocha'
 import {PropTypes} from 'ember-prop-types'
 import {helpers} from 'ember-frost-bunsen/components/button-group-input'
-import {validatePropTypes} from '../../utils/template'
+import {validatePropTypes} from 'dummy/tests/helpers/template'
+import {renderErrorMessageTests} from 'dummy/tests/helpers/abstract-input'
 
 describeComponent(
   'frost-bunsen-input-button-group',
@@ -32,6 +33,7 @@ describeComponent(
       ])
     })
 
+    const ctx = {}
     let component, sandbox
 
     beforeEach(function () {
@@ -51,52 +53,14 @@ describeComponent(
         store: Ember.Object.create({}),
         state: Ember.Object.create({})
       })
+      ctx.component = component
     })
 
     afterEach(function () {
       sandbox.restore()
     })
 
-    describe('renderErrorMessage when error message present', function () {
-      let errorMessage
-
-      beforeEach(function () {
-        errorMessage = 'Things are borked'
-        component.set('errorMessage', errorMessage)
-      })
-
-      describe('when store.showAllErrors is true', function () {
-        beforeEach(function () {
-          component.set('store.showAllErrors', true)
-        })
-
-        it('returns error messages when showErrorMessage is true', function () {
-          component.set('showErrorMessage', true)
-          expect(component.get('renderErrorMessage')).to.eql(errorMessage)
-        })
-
-        it('returns error messages when showErrorMessage is false', function () {
-          component.set('showErrorMessage', false)
-          expect(component.get('renderErrorMessage')).to.eql(errorMessage)
-        })
-      })
-
-      describe('when store.showAllErrors is false', function () {
-        beforeEach(function () {
-          component.set('store.showAllErrors', false)
-        })
-
-        it('returns error messages when showErrorMessage is true', function () {
-          component.set('showErrorMessage', true)
-          expect(component.get('renderErrorMessage')).to.eql(errorMessage)
-        })
-
-        it('returns null when showErrorMessage is false', function () {
-          component.set('showErrorMessage', false)
-          expect(component.get('renderErrorMessage')).to.be.null
-        })
-      })
-    })
+    renderErrorMessageTests(ctx)
 
     describe('options', function () {
       let validateValuesSpy
