+++ tests/unit/components/container-test.js
@@ -1,6 +1,6 @@
 import {describeComponent} from 'ember-mocha'
 import {PropTypes} from 'ember-prop-types'
-import {validatePropTypes} from '../../utils/template'
+import {validatePropTypes} from 'dummy/tests/helpers/template'
 
 describeComponent(
   'frost-bunsen-container',
