+++ tests/unit/components/detail-test.js
@@ -1,7 +1,7 @@
 /* Test fails on continuous integration with 'Symbol is not defined'
 import {describeComponent} from 'ember-mocha'
 import {PropTypes} from 'ember-prop-types'
-import {validatePropTypes} from '../../utils/template'
+import {validatePropTypes} from 'dummy/tests/helpers/template'
 
 describeComponent(
   'frost-bunsen-detail',
