+++ tests/unit/components/static-input-test.js
@@ -2,7 +2,7 @@ import {expect} from 'chai'
 import {describeComponent} from 'ember-mocha'
 import {beforeEach, describe, it} from 'mocha'
 import {PropTypes} from 'ember-prop-types'
-import {validatePropTypes} from '../../utils/template'
+import {validatePropTypes} from 'dummy/tests/helpers/template'
 
 describeComponent(
   'frost-bunsen-input-static',
